fix: change download function
Windows doesn't allow files to be overwritten on the OS if they're in
use, unlike Linux, so we need to drop the new exe in the Downloads dir
for the current user. There could be a better way of doing this but a
simple copy/paste from folder to folder isn't the worst user experience.